:recycle: refactor token system interface
