+++ components/theme/interface/maps/style.ts
@@ -1,0 +1,10 @@
+export interface StyleMapToken {
+  // Line
+  lineWidthBold: number;
+
+  // Radius
+  borderRadiusXS: number;
+  borderRadiusSM: number;
+  borderRadiusLG: number;
+  borderRadiusOuter: number;
+}
